requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-location-field==2.1.0
django-qr-code==1.0.0
djangorestframework==3.10.3
pytz==2019.3
qrcode==6.1
six==1.13.0
sqlparse==0.3.0
